requirements.txt
dill==0.3.1.1
evol==0.5.1
imgcompare==0.1.0
multiprocess==0.70.9
Pillow==5.2.0
py-spy==0.1.8
